update channel history tool
