+++ mcp_servers/slack/bot_tools/__init__.py
@@ -1,13 +1,9 @@
-from .channels import get_channel_history
 from .messages import post_message, reply_to_thread, add_reaction, get_thread_replies
 from .users import get_users, get_user_profile
 from .search import search_messages
 from .base import bot_token_context
 
 __all__ = [
-    # Channels
-    "get_channel_history",
-    
     # Messages
     "post_message",
     "reply_to_thread",
