+++ mcp_servers/slack/user_tools/__init__.py
@@ -1,7 +1,7 @@
 from .user_profile import set_user_status, get_user_profile, set_user_presence
 from .user_search import search_user_messages, search_user_files
 from .direct_messages import open_direct_message, post_direct_message, post_ephemeral_message
-from .channels import list_channels
+from .channels import list_channels, get_channel_history
 from .base import user_token_context
 
 __all__ = [
@@ -21,6 +21,7 @@ __all__ = [
     
     # Channels
     "list_channels",
+    "get_channel_history",
     
     # Base
     "user_token_context",
